requirements.txt
feedparser>=6.0.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
httpx>=0.27.0
google-generativeai>=0.8.0
openai>=1.0.0
python-dateutil>=2.8.0
